requirements.txt
absl-py==0.9.0
astor==0.8.1
autopep8==1.5
click==7.1.1
Flask==1.1.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
numpy==1.18.2
opt-einsum==3.2.0
protobuf==3.11.3
pycodestyle==2.5.0
six==1.14.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
